refactor: exclude the edit hash key tests to a separate suite
re #RI-6570
